+++ RESTART/USER/main.h
@@ -54,9 +54,9 @@
 #include "filter.h"
 #include "kalman_filter.h"
 
-extern int testnum1;		
-extern int testnum2;
-extern int testnum3;
-extern int testnum4;
+extern float testnum1;		
+extern float testnum2;
+extern float testnum3;
+extern float testnum4;
 
 #endif
